May 22/requirements.txt
flask==3.0.2
chromadb==0.4.24
google-generativeai==0.3.2
python-dotenv==1.0.1
waitress==3.0.0
nest-asyncio==1.6.0
numpy==1.26.4
pydantic==1.10.13
tiktoken==0.6.0